refactor: create constant assertions
